Moved some tests in prep for making yet more tests.
