+++ tests/test_asset_exploration.py
@@ -108,16 +108,6 @@ Grid width: 10 pixels.
 """
 
 
-def test_get_asset_choices():
-    """Tests asset choices are correctly pulled out"""
-    assetpack = get_test_assetpack()
-    choices = ddl.asset_exploration.get_asset_choices(assetpack)
-    assert choices[0] == 'Back'
-    assert choices[2] == 'Component: test.a'
-    assert choices[3] == 'Component: test.b'
-    assert choices[5] == 'Image: test.c'
-
-
 def test_print_image_info(capsys):
     """Tests image info prints"""
     image = ImageAsset({"name": 'image',
